moved code to decorator package
