+++ tray/src/com/dvt/main/structural/decorator/BaseDecorator.java
@@ -1,4 +1,4 @@
-package structural;
+package structural.decorator;
 
 public class BaseDecorator implements Person{
     protected Person wrappee;
