+++ tray/src/com/dvt/main/structural/decorator/Person.java
@@ -1,4 +1,4 @@
-package structural;
+package structural.decorator;
 
 public interface Person {
     void wearItem();
