+++ tray/src/com/dvt/main/structural/decorator/Teenager.java
@@ -1,4 +1,4 @@
-package structural;
+package structural.decorator;
 
 public class Teenager implements Person{
     @Override
